+++ CHANGES.md
@@ -93,6 +93,7 @@ v0.7 (unreleased)
   ``glue.qt.mime``                                   | ``glue.core.qt.mime``
   ``glue.qt.qtutil.ComponentIDCombo``                | ``glue.core.qt.component_id_combo``
   ``glue.qt.qtutil.RGBEdit``                         | ``glue.viewers.image.qt.rgb_edit.RGBEdit``
+  ``glue.qt.qtutil.GlueListWidget``              | ``glue.core.qt.mime.GlueMimeListWidget``
 
 * Improved under-the-hood creation of ROIs for Scatter and Histogram Clients. [#676]
