+++ glue/qt/tests/test_qtutil.py
@@ -29,22 +29,3 @@ def test_glue_action_button():
     assert b.text() == 'test2'
 
 
-class TestGlueListWidget(object):
-
-    def setup_method(self, method):
-        self.w = qtutil.GlueListWidget()
-
-    def test_mime_type(self):
-        assert self.w.mimeTypes() == [qtutil.LAYERS_MIME_TYPE]
-
-    def test_mime_data(self):
-        self.w.set_data(3, 'test data')
-        self.w.set_data(4, 'do not pick')
-        mime = self.w.mimeData([3])
-        mime.data(qtutil.LAYERS_MIME_TYPE) == ['test data']
-
-    def test_mime_data_multiselect(self):
-        self.w.set_data(3, 'test data')
-        self.w.set_data(4, 'also pick')
-        mime = self.w.mimeData([3, 4])
-        mime.data(qtutil.LAYERS_MIME_TYPE) == ['test data', 'also pick']
